# [Xamarin.Android.Build.Tasks] `FixAbstractMethodsStep` performance