@jonathanpeppers 2024-01-19
Ignore the MAUI lane for now, I think their build is broken:
```
C:\a\_work\3\s\maui\src\Core\src\Platform\Tizen\StackNavigationManager.cs(73,25): error CA1859: Change type of variable 'newPageStack' from 'System.Collections.Generic.IReadOnlyList<Microsoft.Maui.IView>' to 'System.Collections.Generic.List<Microsoft.Maui.IView>' for improved performance (https://learn.microsoft.com/dotnet/fundamentals/code-analysis/quality-rules/ca1859) [C:\a\_work\3\s\maui\src\Core\src\Core.csproj::TargetFramework=net9.0-tizen]
C:\a\_work\3\s\maui\src\Core\src\Platform\Windows\StackNavigationManager.cs(65,25): error CA1859: Change type of variable 'newPageStack' from 'System.Collections.Generic.IReadOnlyList<Microsoft.Maui.IView>' to 'System.Collections.Generic.List<Microsoft.Maui.IView>' for improved performance (https://learn.microsoft.com/dotnet/fundamentals/code-analysis/quality-rules/ca1859) [C:\a\_work\3\s\maui\src\Core\src\Core.csproj::TargetFramework=net9.0-windows10.0.20348]
C:\a\_work\3\s\maui\src\Core\src\Platform\Windows\StackNavigationManager.cs(65,25): error CA1859: Change type of variable 'newPageStack' from 'System.Collections.Generic.IReadOnlyList<Microsoft.Maui.IView>' to 'System.Collections.Generic.List<Microsoft.Maui.IView>' for improved performance (https://learn.microsoft.com/dotnet/fundamentals/code-analysis/quality-rules/ca1859) [C:\a\_work\3\s\maui\src\Core\src\Core.csproj::TargetFramework=net9.0-windows10.0.19041]
```